App/requirements.txt
pandas
pickle-mixin
nltk
sklearn
requests
flask
python-docx